+++ spec/spec_runner.js
@@ -14,4 +14,3 @@
       jasmineMajorVersion: 2
    };
 } )( this );
-