+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Last Changes
 
-- [#10](https://github.com/LaxarJS/ax-command-bar-widget/issues/10): Use laxar-mocks instead of laxar-testing now.
+- [#10](https://github.com/LaxarJS/ax-command-bar-widget/issues/10): Use laxar-mocks instead of laxar-testing now
 
 
 ## v2.3.0